baseline_dvbf/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.12.0=py36h06a4308_0
astor=0.8.1=py36h06a4308_0
blas=1.0=mkl
c-ares=1.17.1=h27cfd23_0
ca-certificates=2022.4.26=h06a4308_0
certifi=2021.5.30=py36h06a4308_0
cloudpickle=1.6.0=pypi_0
coverage=5.5=py36h27cfd23_2
cudatoolkit=9.2=0
cudnn=7.6.5=cuda9.2_0
cupti=9.2.148=0
cycler=0.10.0=pypi_0
cython=0.29.23=py36h2531618_0
dbus=1.13.18=hb2f20db_0
expat=2.4.4=h295c915_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
gast=0.4.0=py_0
glib=2.69.1=h4ff587b_1
grpcio=1.36.1=py36h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gym=0.18.3=pypi_0
h5py=2.10.0=py36hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
importlib-metadata=3.10.0=py36h06a4308_0
intel-openmp=2021.2.0=h06a4308_610
jpeg=9e=h7f8727e_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py36h06a4308_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1o=h7f8727e_0
pandas=1.1.5=py36ha9443f7_0
pcre=8.45=h295c915_0
pillow=8.2.0=pypi_0
pip=21.1.1=py36h06a4308_0
progressbar=2.5=pypi_0
protobuf=3.14.0=py36h2531618_1
pyglet=1.5.15=pypi_0
pyparsing=2.4.7=pypi_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=hdb3f193_0
python-dateutil=2.8.1=pypi_0
pytz=2021.3=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
scipy=1.5.2=py36h0b6359f_0
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=52.0.0=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py36h06a4308_0
sqlite=3.35.4=hdfb4753_0
tensorboard=1.11.0=py36hf484d3e_0
tensorflow=1.11.0=gpu_py36h9c9050a_0
tensorflow-base=1.11.0=gpu_py36had579c0_0
tensorflow-gpu=1.11.0=h0d30ee6_0
termcolor=1.1.0=py36h06a4308_1
tk=8.6.10=hbc83047_0
tornado=6.1=py36h27cfd23_0
typing_extensions=3.7.4.3=pyha847dfd_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
